requirements.txt
groq==0.4.2
streamlit==1.32.0
python-dotenv==1.0.1
httpx==0.26.0